external-hdf: find XSA files in a folder
Previously each variant would be defined in PL_VARIANTS and the XSA
files placed into a folder local to the recipe.

With this change, only the folder location PL_VARIANT_DIR is specified.
In the simplest case this is the recipe folder, e.g.:

  PL_VARIANTS_DIR := "${THISDIR}/${MACHINE}"

It adds Python code that runs before the script functions are called and
recursively globs for all XSA files in it. The SRC_URI and FILES
variables are updated accordingly. Changes to the XSA files should be
noticed by bitbake so it rebuilds everything.

Alternatively one can still specify a single XSA file with HDF_PATH.

The way XSA files are installed into the sysroot should not have
changed. The `pl-variants` file still gets populated.

Code was partly taken from Xilinx' `fpga-manager-util_1.0.bb` recipe
which does a very similar thing.